requirements.txt
datasets>=2.18.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
sentence-transformers>=2.7.0
openai>=1.30.0
tiktoken>=0.7.0
numpy>=1.26.0
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
pandas>=2.2.0
rank-bm25>=0.2.2
